moved fulfilment details into the header on desktops
