+++ js/app/restaurant/header.scss
@@ -16,11 +16,11 @@
 .restaurant-header {
   display: flex;
   flex-direction: column;
+  padding: 1.5rem 0;
 
   &-body {
     display: flex;
     flex-direction: column;
-    padding-bottom: 1.5rem;
 
     @media (min-width: $screen-md-min) {
       flex-direction: row;
@@ -33,14 +33,14 @@
     display: flex;
     flex-direction: column;
     gap: 1rem;
-  }
 
-  a {
-    color: $text-color;
-    &:hover,
-    &:focus {
-      background-color: $light-gray;
-      text-decoration: none;
+    a {
+      color: $text-color;
+      &:hover,
+      &:focus {
+        background-color: $light-gray;
+        text-decoration: none;
+      }
     }
   }
 
@@ -73,7 +73,7 @@
   }
 
   .restaurant-name {
-    margin-bottom: 0;
+    margin: 0;
     font-weight: bold;
   }
